Desktop/projects/ANTIGRAVITY_NEUROVAULT/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
pydantic==2.7.1
pydantic-settings==2.2.1
python-multipart==0.0.9
aiofiles==23.2.1
numpy
sentence-transformers
faiss-cpu
PyPDF2==3.0.1
python-docx==1.1.0
Pillow
pytesseract
scikit-learn
httpx
